Move AddToken into layouts folder
