+++ src/layouts/MainLayout.tsx
@@ -21,7 +21,7 @@ import { CHAIN_NAMES, ChainId } from '../library/constants/chains';
 import { CHAIN_INFO } from '../configs';
 
 import { ColorModeSwitcher } from '../components/ColorModeSwitcher';
-import AddToken from '../components/AddToken';
+import AddToken from './AddToken';
 
 import { ReactComponent as SvgMenuOpen } from '../assets/menu-open.svg';
 import { ReactComponent as SvgMenuClose } from '../assets/menu-close.svg';
